backend/requirements.txt
Django==3.2.16
django-extra-fields
django-filter==21.1
djangorestframework==3.12.4
djoser==2.1.0
gunicorn==20.1.0
Pillow==9.3.0
psycopg2-binary
pyparsing==3.0.9
python-dotenv==0.19.2
sqlparse==0.4.3